+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - [TD-5289] Elasticsearch 7 compatibility
 - [TD-5342] Refactor bulk operations to use Elasticsearch scroll API
 
+### Added
+
+- [TD-5258] Approve grants request rules
+
 ## [4.55.0] 2022-11-14
 
 ### Added